added json collision loading, fixed mem leaks

+++ Application/constants.hpp
@@ -1,7 +1,7 @@
 #pragma once
 namespace constants
 {
-	const int SCREEN_WIDTH{ 900 };
-	const int SCREEN_HEIGHT{ 600 };
+	const float SCREEN_WIDTH{ 900 };
+	const float SCREEN_HEIGHT{ 600 };
 }
